requirements.txt
amqp==1.4.9
anyjson==0.3.3
asn1crypto==0.22.0
bcrypt==3.1.3
billiard==3.3.0.23
celery==3.1.20
cffi==1.10.0
cryptography==1.9
Django==1.8.8
django-admin-bootstrapped==2.5.7
django-allauth==0.19.1
django-appconf==1.0.1
django-bootstrap3==7.1.0
django-bower==5.1.0
django-celery==3.1.17
django-cors-headers==1.2.0
django-fields==0.3.0
django-nose==1.4.3
django-snippetscream==0.0.7
django-storages==1.1.8
django-tastypie==0.13.0
django-test-without-migrations==0.4
enum34==1.1.6
Fabric==1.10.2
factory-boy==2.4.0
fake-factory==0.5.3
httplib2==0.9.2
idna==2.5
ipaddr==2.1.11
ipaddress==1.0.16
jsonfield==1.0.3
kombu==3.0.33
longerusername==0.4
mock==1.0.1
MySQL-python==1.2.5
nose==1.3.7
numpy==1.11.0
oauth2client==1.5.2
oauthlib==1.0.3
olefile==0.44
paramiko==2.2.1
Pillow==4.2.1
pyasn1==0.2.3
pyasn1-modules==0.0.9
pycparser==2.18
pycrypto==2.6.1
PyNaCl==1.1.2
python-dateutil==2.4.2
python-mimeparse==1.6.0
python-openid==2.2.5
pytz==2015.7
reportlab==3.2.0
requests==2.9.1
requests-mock==0.7.0
requests-oauthlib==0.6.0
rsa==3.3
simplejson==3.8.1
six==1.10.0
ua-parser==0.6.1
user-agents==1.0.1
virtualenv==14.0.1
